pyproject.toml
[tool.poetry]
name = "big-project"
version = "0.1.0"
description = ""
authors = ["alexander anciferov <sania.antsiferoff@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.1.0"}
django-cors-headers = "^4.7.0"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
